+++ packages/toolkit/config/filenames.config.js
@@ -2,6 +2,6 @@ module.exports = {
 	js: 'js/[name].js',
 	jsChunk: 'js/[name].[contenthash].chunk.js',
 	css: 'css/[name].css',
-	block: 'blocks/[name]/editor.js',
-	blockCSS: 'blocks/[name]/editor.css',
+	block: 'blocks/[name].js',
+	blockCSS: 'blocks/[name].css',
 };