+++ packages/toolkit/config/paths.config.js
@@ -2,4 +2,5 @@ module.exports = {
 	srcDir: './assets/',
 	cssLoaderPaths: ['./assets/css', './includes/blocks'],
 	copyAssetsDir: './assets/',
+	blocksDir: './includes/blocks/',
 };