+++ packages/toolkit/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 4.2.0-next.2
+
+### Minor Changes
+
+- 2e67b06: Replaces `squoosh` with a custom implementation using `sharp` for optimizing images.
+
+## 4.2.0-next.1
+
+### Patch Changes
+
+- bc89638: Fix how webpack handles addition of new block.json files during watch mode
+
+## 4.2.0-next.0
+
+### Minor Changes
+
+- d60ce6c: Refine the way block assets get handled. 10up-toolkit will now create Webpack entrypoints for any assets that are defined in any block.json files automatically for you. So no need to manually adding manual entrypoints per block.
+
 ## 4.1.2
 
 ### Patch Changes